+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.4.1
+
+* Fix skewed ecliptic wheel angles than had been baked into the original SVG graphics.
+
+## 1.4.0
+
+* New option for a speculated original 1410 appearance of the orloj.
+
 ## 1.3.3
 
 * Added ecliptic longitude tooltips to the large sun and moon icons, instead of just the smaller true positions markers.